// src/model/Branch.ts
// Generated by https://quicktype.io

export interface Branch {
    id:                number;
    name:              string;
    description:       null | string;
    branch:            null;
    delivery_time:     null | string;
    address:           null | string;
    phone:             null | string;
    schedule:          null | string;
    logo:              string;
    banner:            string;
    location:          Location;
    distance:          string;
    rating:            string;
    allow_cash:        boolean;
    distance_for_free: string;
    amount_for_free:   null;
}

export interface Location {
    lat: number;
    lng: number;
}